Revert "Enabling c++ 20 in MacOS build (#16187)"
This reverts commit 0141e27ca1e9954da2630f817709e2eebcd88dde.

It is unclear to me why C++20 was enabled specifically for macOS.
The git revision log does not state why the change was made but
based on the diff it looks like it was done mainly so that
unicode prefixes, "u8", from string literals.

The motivation to dropping macOS back down to C++17 (like Linux
and Windows) is to fix build errors from date/date.h C++20
compatibility.  There's a fix for these errors upstream in the
main branch, 528b8a5410366198150117a97066b98353b3bd00, but
applying the fix produces new errors, "'to_chars' is unavailable,
introduced in macOS 13.3".  This occurs because we currently
build with a min. macOS target version of less than 13.3.
We cannot drop support for macOS versions < 13.3 at this time
so the only solution is to drop macOS back down to C++17.